Refactor: apply more oop style architecture
Before style architecture is still mixed with procedure and oop style.
Currently, we only have very poor test, so i think that one big change
is more better than gradual refactoring.

I reference the following link and came to the following conclusion.
(https://stackoverflow.com/questions/325453/how-to-write-main-in-an-oop-way)

Main function to be simple and made up of a combination of objects.
It also becomes easier to test by dividing it into individual objects.